requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
httpx==0.28.1
jinja2==3.1.4
structlog==24.4.0
numpy
pyinstaller
pyyaml
anthropic>=0.52.0
mcp[server]>=1.0.0
sse-starlette>=2.0.0
